+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 1. ES显示文档内容，长整型精度丢失问题修复
 2. ES模块数据根据字段每个列单独展示
 3. ES添加字段搜索、排序功能
+4. Redis、Zookeeper等JSON字符串预览
+5. Redis的Hash、Set、List的值如果是JSON字符串，添加预览按钮
 
 # 1.5.7